+++ ducatus_exchange/transfers/api.py
@@ -1,21 +1,25 @@
-# import time
+import time
 import sys
 import traceback
 import logging
 from decimal import Decimal
 
 from ducatus_exchange.litecoin_rpc import DucatuscoreInterface
-# from ducatus_exchange.parity_interface import ParityInterface
+from ducatus_exchange.parity_interface import ParityInterface, ParityInterfaceException
 from ducatus_exchange.transfers.models import DucatusTransfer
 from ducatus_exchange.settings import ROOT_KEYS, REF_BONUS_PERCENT, MINIMAL_RETURN
 from ducatus_exchange.bip32_ducatus import DucatusWallet
-# from ducatus_exchange.consts import DAYLY_LIMIT, WEEKLY_LIMIT
-# from ducatus_exchange.payments.utils import calculate_amount
-# from ducatus_exchange.exchange_requests.models import ExchangeRequest
+from ducatus_exchange.consts import DAYLY_LIMIT, WEEKLY_LIMIT
+from ducatus_exchange.payments.utils import calculate_amount
+from ducatus_exchange.exchange_requests.models import ExchangeRequest
 from ducatus_exchange.ducatus_api import return_ducatus, return_ducatusx
-# from ducatus_exchange.exchange_requests.models import ExchangeStatus
+from ducatus_exchange.exchange_requests.models import ExchangeStatus
 
 logger = logging.getLogger(__name__)
+
+
+class TransferException(Exception):
+    pass
 
 
 def save_transfer(payment, tx, amount, currency):
@@ -32,6 +36,38 @@ def save_transfer(payment, tx, amount, currency):
 
     logger.info(msg='transfer saved')
     return transfer
+
+
+def check_limits(payment):
+    dayly_reserve = 0
+    weekly_reserve = 0
+    original_amount=payment.original_amount
+    if payment.original_amount + payment.exchange_request.dayly_swap > DAYLY_LIMIT:
+        dayly_reserve = DAYLY_LIMIT - payment.exchange_request.dayly_swap
+        logger.info(msg=dayly_reserve)
+        payment.original_amount = dayly_reserve
+        if dayly_reserve <= 0:
+            return False, original_amount
+    if payment.original_amount + payment.exchange_request.weekly_swap > WEEKLY_LIMIT:
+        weekly_reserve = WEEKLY_LIMIT - payment.exchange_request.weekly_swap
+        if dayly_reserve > 0:
+            payment.original_amount = min(dayly_reserve, weekly_reserve)
+        else:
+            payment.original_amount = weekly_reserve
+        if weekly_reserve <= 0:
+            return False, original_amount
+    logger.info(msg=f' amount {payment.original_amount}')
+    exchange_request=ExchangeRequest.objects.get(id=payment.exchange_request.id)
+    exchange_request.dayly_swap += payment.original_amount
+    exchange_request.weekly_swap += payment.original_amount
+    exchange_request.save()
+    logger.info(msg=f'daily {exchange_request.dayly_swap}')
+    if payment.original_amount !=original_amount:
+        payment.sent_amount, payment.rate = calculate_amount(payment.original_amount, payment.currency)
+        logger.info(msg=f"User's {payment.exchange_request.user.id} sent_amount was recalculated due to limits")
+        payment.save()
+        return True, original_amount-payment.original_amount
+    return True, 0
 
 
 def make_ref_transfer(payment):
@@ -71,29 +107,52 @@ def transfer_ducatus(payment):
         return None
 
 
-# def transfer_ducatusx(payment):
-#     amount = payment.sent_amount
-#     receiver = payment.exchange_request.user.address
-#     logger.info(msg=f'ducatusX transfer started: sending {amount} DUCX to {receiver}')
-#     currency = 'DUCX'
+def add_transfer_duc_in_queue(payment):
+    payment.state_transfer_in_queue()
+    payment.save()
 
-#     parity = ParityInterface()
-#     # if not enough balance on admin address return tokens to user
-#     if parity.get_balance() > amount:
-#         tx = parity.transfer(receiver, amount)
-#         transfer = save_transfer(payment, tx, amount, currency)
 
-#         logger.info(msg='ducatusx transfer ok')
+def transfer_ducx(payment):
+    try:
+        status = ExchangeStatus.objects.all().first().status
+        if not status:
+            logger.info(msg='exchange is disabled')
+            add_transfer_duc_in_queue(payment) # second task if in queue
+            # return_ducatus(payment.tx_hash, payment.original_amount)
+        else:
+            allowed, return_amount = check_limits(payment)
+            if return_amount > MINIMAL_RETURN:
+                add_transfer_duc_in_queue(payment) # third task if in queue
+                # return_ducatus(payment.tx_hash, return_amount)
+            if allowed:
+                amount = payment.sent_amount
+                receiver = payment.exchange_request.user.address
+                logger.info(msg=f'ducatusX transfer started: sending {amount} DUCX to {receiver}')
+                currency = 'DUCX'
+                parity = ParityInterface()
+                # if not enough balance on admin address return tokens to user
+                if parity.get_balance() > amount:
+                    tx = parity.transfer(receiver, amount)
+                    transfer = save_transfer(payment, tx, amount, currency)
 
-#         time.sleep(100)    # small timeout in case of multiple payment messages
-#         return transfer
-#     else:
-#         logger.info(msg=f'Not enough balance on wallet DUC, transaction with hash {payment.tx_hash} will return to user on DUCX')
-#         return_ducatus(
-#             payment_hash=payment.tx_hash,
-#             amount=amount,
-#         )
-#         return None
+                    logger.info(msg='ducatusx transfer ok')
+
+                    time.sleep(100)    # small timeout in case of multiple payment messages
+                    return transfer
+                else:
+                    logger.info(msg=f'Not enough balance on wallet DUC, transaction with hash {payment.tx_hash} will return to user on DUCX')
+                    add_transfer_duc_in_queue(payment) # 4 test case in queue
+                    # return_ducatus(payment_hash=payment.tx_hash,amount=amount,)
+            else:
+                logger.info(
+                    msg=f"User's {payment.exchange_request.user.id} swap amount reached limits, cancelling transfer"
+                )
+    except ParityInterfaceException as e:
+        payment.state_transfer_error()
+        payment.save()
+        raise TransferException(e)
+    logger.info(msg='transfer completed')
+
 
 
 def confirm_transfer(message):
